+++ R/AllGenerics.R
@@ -63,7 +63,6 @@ setGeneric("self", function (x) standardGeneric("self"))
 setGeneric("refresh", function (x) standardGeneric("refresh"))
 setGeneric("delete", function (x, ...) standardGeneric("delete"),
     signature="x")
-setGeneric("readonly<-", function (x, value) standardGeneric("readonly<-"))
 setGeneric("entities", function (x, ...) standardGeneric("entities"))
 setGeneric("entities<-", function (x, value) standardGeneric("entities<-"))
 setGeneric("tuple", function (x) standardGeneric("tuple"))