+++ tests/testthat/test-get-datasets.R
@@ -2,12 +2,6 @@ context("Retrieving dataset list and single datasets")
 
 with_mock_HTTP({
     cr <- session()
-
-    dataset.catalog.url <- "/api/datasets.json"
-    datcat <- DatasetCatalog(crGET(dataset.catalog.url))
-    index(datcat)[[which(names(datcat) == "an archived dataset")]]$archived <- TRUE
-    session_store$datasets <- datcat ## as if we had done updateDatasetList
-
     test_that("listDatasets lists", {
         expect_identical(listDatasets(), c("ECON.sav", "test ds"))
         expect_identical(listDatasets("archived"), "an archived dataset")
@@ -43,12 +37,6 @@ with_mock_HTTP({
 })
 
 if (run.integration.tests) {
-    test_that("updateDatasetList requires authentication", {
-        logout()
-        expect_error(updateDatasetList(),
-            "You must authenticate before making this request")
-    })
-
     with_test_authentication({
         test_that("datasetCatalog gets what we expect", {
             col0 <- datasetCatalog()