+++ NEWS.md
@@ -1,4 +1,5 @@
 ### crunch 1.10.5 (under development)
+* (Re-)improve print method for expressions involving categorical variables
 
 ### crunch 1.10.4
 * Fix `as.data.frame`/`as.environment` for `CrunchDataset` when a variable alias contained an apostrophe.