+++ NEWS.md
@@ -1,3 +1,5 @@
+### crunch 1.10.7 (under development)
+
 ### crunch 1.10.6
 * Fix refresh method for Datasets that have been transferred to a Project.
 * (Re-)improve print method for expressions involving categorical variables